# Conflicts:
#   app/(content)/browse/page.tsx

+++ lib/db/reviews/base-reviews.ts
@@ -2,10 +2,11 @@
 
 import { hasPermission } from "@/lib/permissions";
 import { prisma } from "@/prisma";
+import { forbidden, PoliError, unauthorized } from "@/utils/misc";
 import { getCurrentUser } from "@/utils/users";
 import { BaseReview, Image, User } from "@prisma/client";
-import { forbidden, unauthorized } from "next/navigation";
 import { deleteImages, getImagesByPaths } from "../images";
+import { getResponseReviewById } from "./response-reviews";
 import {
   getRestaurantReviewById,
   updateRestaurantStats,
@@ -20,7 +21,7 @@ export type BaseReviewFull = {
 
 export async function getBaseReviewById(id: BaseReview["id"]) {
   const currentUser = await getCurrentUser();
-  if (currentUser == null) unauthorized();
+  if (currentUser == null) return unauthorized();
   const baseReview = await prisma.baseReview.findFirst({
     where: {
       id: id,
@@ -35,15 +36,17 @@ export async function getBaseReviewById(id: BaseReview["id"]) {
     baseReview?.hidden &&
     !hasPermission(currentUser, "reviews", "viewHidden", baseReview)
   )
-    forbidden();
+    return forbidden();
   return baseReview;
 }
 
 export async function hideReview(id: BaseReview["id"], hide: boolean) {
   const baseReview = (await getBaseReviewById(id)) || undefined;
+  if (baseReview instanceof PoliError) return baseReview;
   const currentUser = await getCurrentUser();
-  if (currentUser == null) unauthorized();
-  if (!hasPermission(currentUser, "reviews", "hide", baseReview)) forbidden();
+  if (currentUser == null) return unauthorized();
+  if (!hasPermission(currentUser, "reviews", "hide", baseReview))
+    return forbidden();
   return await prisma.baseReview.update({
     where: {
       id: id,
@@ -56,13 +59,18 @@ export async function hideReview(id: BaseReview["id"], hide: boolean) {
 
 export async function deleteReview(id: BaseReview["id"]) {
   const baseReview = (await getBaseReviewById(id)) || undefined;
+  if (baseReview instanceof PoliError) return baseReview;
   if (!baseReview) return null;
   const currentUser = await getCurrentUser();
-  if (currentUser == null) unauthorized();
-  if (!hasPermission(currentUser, "reviews", "delete", baseReview)) forbidden();
+  if (currentUser == null) return unauthorized();
+  if (!hasPermission(currentUser, "reviews", "delete", baseReview))
+    return forbidden();
   await deleteImages(baseReview.images.map((image) => image.path));
   // TODO: find a cleaner way to update restaurant stats
   const restaurantReview = await getRestaurantReviewById(id);
+  if (restaurantReview instanceof PoliError) return baseReview;
+  const responseReview = await getResponseReviewById(id);
+  if (responseReview instanceof PoliError) return baseReview;
   const review = await prisma.baseReview.delete({
     where: {
       id: id,
@@ -70,6 +78,18 @@ export async function deleteReview(id: BaseReview["id"]) {
   });
   if (restaurantReview) {
     await updateRestaurantStats(restaurantReview.subjectId);
+  }
+  if (responseReview) {
+    await prisma.baseReview.update({
+      where: {
+        id: responseReview.subjectId,
+      },
+      data: {
+        responsesAmount: {
+          decrement: 1,
+        },
+      },
+    });
   }
   return review;
 }
@@ -80,14 +100,14 @@ export async function linkImagesToReview(
 ) {
   // permission check
   const user = await getCurrentUser();
-  if (user == null) unauthorized();
+  if (user == null) return unauthorized();
   const images = await getImagesByPaths(imagePaths);
   if (
     !images.every((image) => {
       return hasPermission(user, "images", "delete", image);
     })
   )
-    forbidden();
+    return forbidden();
 
   // Try to connect
   return await prisma.baseReview.update({
